Corrigido arquivos.sh e atualizado leiame.txt

+++ yml/ruby2.2.1_rails4.2.1/down.sh
@@ -1,1 +1,8 @@
-docker-compose down
+#!/bin/bash
+PROJETO="/home/kdepaulo/docker/yml/ruby2.2.1_rails4.2.1/myapp"
+
+if [ "$(pwd)" == $PROJETO ]; then
+    docker-compose down
+else
+    echo "Execute o script $0 da pasta $PROJETO"
+fi
